Sphere now moves to show where you supposedly clicked
git-svn-id: https://paroxysm.googlecode.com/svn/trunk@75 1a4693bc-ef48-11de-b206-85b67c77eac7

+++ source/TSphere.cpp
@@ -63,7 +63,7 @@ void TSphere::display()
     glPopAttrib();
 }
 
-void TSphere::moveSphere(int inX, int inY, int inZ)
+void TSphere::moveSphere(float inX, float inY, float inZ)
 {
     mTranslation[0] = inX;
     mTranslation[1] = inY;
